"Update filters and table display in Clients page"
This commit includes updates to the filtering and display of data in the Clients page. The width of the menu has been increased from 130px to 140px. The Select component has been updated to use the `useMemo` hook for filtering by date, and the `dateFilter` action has been added to the `roleUsers` slice in the Redux store. The FilterData component has also been updated to use the `useDispatch` hook and the `filterUsers` action. The Table component has been updated to display a message when no data is found. Additionally, the `FilterTable` component's outerStyle transition time has been increased to 0.5s. The `menu.module.css` file has also been updated to change the width of the list to 140px.